maintain song order in album view
